requirements.txt
streamlit==1.38.0
pydantic==2.8.2
loguru==0.7.2
pdfplumber==0.11.4
python-docx==1.1.2
spacy==3.7.4
sentence-transformers==3.0.1
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
plotly==5.23.0
rapidfuzz==3.9.6
trafilatura==1.9.0
duckduckgo_search==6.2.7
tqdm==4.66.4
joblib==1.4.2
requests==2.32.3
